Point people filing issues to discussions
We are not yet at the point where Moonwalk will be managed
through issues.  This changes the new issue workflow
to offer buttons pointing to GitHub discussions and/or Slack,
along with a template that repeats the advice from #11 to open
a discussion instead of an issue.

I somewhat strengthened #11's language with the criteria that
issues need to be things we can resolve by PRs in the short term.